alx_travel_app/requirement.txt
Django>=4.0
djangorestframework
django-cors-headers
drf-yasg
django-environ
mysqlclient
celery
pika
